requirements.txt
pandas==2.2.3
numpy==2.2.6
scikit-learn==1.6.1
ipykernel
matplotlib==3.10.3
plotly==6.1.2
openai==1.83.0
faiss-cpu==1.11.0
torch==2.7.0
torchvision==0.22.0
transformers==4.52.4
open_clip_torch==2.32.0
sentence-transformers==4.1.0
streamlit==1.45.1